Keep the MAXBYTES threshold 64-bit wide and extend the tests
streamTrim() narrowed the parsed MAXBYTES threshold to size_t, so on a
32-bit build a threshold of 4294967296 became 0 and the trim removed
every node. Keep the threshold uint64_t; total_lp_bytes stays size_t
since it only ever holds sums of lpBytes().

Tests added:
* Thresholds of 2^32 and 2^32+128 trim nothing on XTRIM and XADD.
* A non-aligned threshold with an oversized head node stops before the
  node whose removal would undershoot.
* Interleaved appends of different layouts, XDEL, exact MAXLEN/MINID and
  MAXBYTES keep the tracked bytes equal to a recount after every step.
* MAXBYTES trims replay from the incremental and the rewritten AOF under
  a different node size, and the rebuilt counter trims again.
* A replica with different stream-node-max-entries/-bytes holds the
  same entries after trimming and non-trimming XADD, LIMIT and XTRIM.
* Parser rejects ~, =, a duplicate MAXBYTES, overflow and negative LIMIT.

+++ src/t_stream.c
@@ -721,7 +721,7 @@ typedef struct {
 int64_t streamTrim(stream *s, streamAddTrimArgs *args) {
     size_t maxlen = args->maxlen;
     streamID *id = &args->minid;
-    size_t maxbytes = args->maxbytes;
+    uint64_t maxbytes = (uint64_t)args->maxbytes;
     int approx = args->approx_trim;
     int64_t limit = args->limit;
     int trim_strategy = args->trim_strategy;
